// cosmowcl/src/app/shared/models/api/cosmoauthorization/fruitore.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cw-profilazione
 * API dei servizi di profilazione usati dal common workspace
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EndpointFruitore } from './endpointFruitore.d';
import { SchemaAutenticazioneFruitore } from './schemaAutenticazioneFruitore.d';
import { AssociazioneFruitoreEnte } from './associazioneFruitoreEnte.d';
import { AutorizzazioneFruitore } from './autorizzazioneFruitore.d';


export interface Fruitore { 
    id?: number;
    apiManagerId: string;
    url?: string;
    nomeApp?: string;
    enti?: Array<AssociazioneFruitoreEnte>;
    autorizzazioni?: Array<AutorizzazioneFruitore>;
    endpoints?: Array<EndpointFruitore>;
    schemiAutenticazione?: Array<SchemaAutenticazioneFruitore>;
}